more styling/refactoring
