+++ src/components/right-panel/scorecardFull.jsx
@@ -1,0 +1,11 @@
+import React, { useState, useEffect } from 'react';
+import { MainScore } from "./scorecardMain";
+import { connect } from "react-redux";
+
+
+export const FullScorecard = () => {
+    
+    return (
+        <MainScore />
+    );
+}
